src/chef/continue_reorganize.py
#!/usr/bin/env python3
"""
Videous Project Reorganizer - Continuation

This script continues the reorganization process after fixing the model file issue.
"""

import os
import shutil
import re
import glob
import configparser
import sys
import traceback
from datetime import datetime

# Check for required dependencies
try:
    import yaml
    from pathlib import Path
    print("All required dependencies are installed.")
except ImportError as e:
    print(f"Error: Missing dependency - {e}")
    print("Please install required dependencies: pip install pyyaml")
    sys.exit(1)

# Import the functions from the original script
from reorganize_project import (
    convert_config_to_yaml,
    create_package_structure,
    clean_up_redundant_files,
    create_documentation,
    create_requirements_file,
    create_gitignore,
    update_import_statements,
    create_s3_integration_module,
    create_config_manager,
    create_setup_script,
    create_test_structure
)

def continue_reorganization():
    """Continue the reorganization process."""
    try:
        # Get the base directory (current directory)
        base_dir = os.getcwd()
        
        print(f"Script starting at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
        print(f"Current working directory: {base_dir}")
        print(f"Continuing Videous project reorganization in: {base_dir}")
        print("=" * 50)
        
        # Execute remaining reorganization steps
        convert_config_to_yaml(base_dir)
        create_package_structure(base_dir)
        clean_up_redundant_files(base_dir)
        create_documentation(base_dir)
        create_requirements_file(base_dir)
        create_gitignore(base_dir)
        
        # Additional functionality
        update_import_statements(base_dir)
        create_s3_integration_module(base_dir)
        create_config_manager(base_dir)
        create_setup_script(base_dir)
        create_test_structure(base_dir)
        
        print("\n" + "=" * 50)
        print("Project reorganization complete!")
        print("\nNext steps:")
        print("1. Review the new structure and make any necessary adjustments")
        print("2. Update import statements in Python files")
        print("3. Test the reorganized codebase")
        print("4. Commit the changes to version control")
        print(f"\nScript completed at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
        
    except Exception as e:
        print(f"\nError during reorganization: {e}")
        print("\nStack trace:")
        traceback.print_exc()
        print("\nReorganization failed. Please check the error message above.")
        sys.exit(1)

if __name__ == "__main__":
    continue_reorganization()
