requirements.txt
# Consolidated requirements for Videous & Friends project

Cython==3.0.12
Jinja2==3.1.4
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==2.1.5
MarkupSafe==3.0.2
Pillow==9.2.0
PyYAML==6.0.2
Pygments==2.19.1
Werkzeug==3.1.3
absl-py==1.2.0
absl-py==2.2.1
altgraph==0.17.2
astunparse==1.6.3
attrs==22.1.0
attrs==25.3.0
audioread==3.0.1
autopep8==2.3.2
av==14.3.0
black==25.1.0
cachetools==5.5.2
certifi==2022.6.15
certifi==2024.8.30
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==2.1.1
charset-normalizer==3.3.2
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
colour-science==0.4.6
contourpy==1.3.1
customtkinter==4.5.10
customtkinter==5.2.2
cycler==0.11.0
cycler==0.12.1
darkdetect==0.7.1
darkdetect==0.8.0
decorator==4.4.2
distlib==0.3.9
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
filelock==3.18.0
filterpy==1.4.5
flake8==7.2.0
flatbuffers==25.2.10
fonttools==4.37.1
fonttools==4.56.0
fsspec==2024.9.0
fsspec==2025.3.2
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth-oauthlib==1.0.0
google-auth==2.39.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.1
identify==2.6.9
idna==3.10
idna==3.3
imageio-ffmpeg==0.5.1
imageio==2.35.1
iopath==0.1.10
jax==0.4.30
jaxlib==0.4.30
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.4
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib
matplotlib==3.10.1
matplotlib==3.5.3
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.21
mediapipe==0.8.10.1
mido==1.3.3
ml-dtypes==0.3.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
networkx==3.4.2
nodeenv==1.9.1
numba==0.60.0
numpy
numpy==1.21.2
numpy==1.23.5
numpy==2.0.2
oauthlib==3.2.2
openai-whisper==20231117
opencv-contrib-python==4.11.0.86
opencv-contrib-python==4.6.0.66
opencv-python
opencv-python==4.11.0.86
opencv-python==4.5.5.64
opt_einsum==3.4.0
optree==0.15.0
packaging==21.3
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
pefile==2022.5.30
pillow
pillow==10.4.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
portalocker==3.1.1
pre_commit==4.2.0
pretty_midi==0.2.10
proglog==0.1.10
protobuf==3.20.1
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
pyinstaller @ https://github.com/pyinstaller/pyinstaller/tarball/develop
pyinstaller-hooks-contrib==2022.9
pyparsing==3.0.9
pyparsing==3.2.3
python-dateutil==2.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorchvideo==0.1.5
pytz==2025.2
pywin32-ctypes==0.2.0
regex==2024.9.11
reindent==3.5.1
requests-oauthlib==2.0.0
requests==2.28.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.5
safetensors==0.5.3
scikit-image
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentencepiece==0.2.0
six==1.16.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
sympy==1.13.2
tabulate==0.9.0
tensorboard-data-server==0.7.2
tensorboard==2.12.3
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.12.0
tensorflow-metal==1.2.0
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.15
tkinterdnd2==0.4.3
tokenizers==0.21.1
tomli==2.2.1
torch==1.13.1
torch==2.4.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm
tqdm==4.66.5
tqdm==4.67.1
transformers==4.50.2
typing_extensions==4.12.2
typing_extensions==4.13.0
tzdata==2025.2
ultralytics-thop==2.0.14
ultralytics==8.3.99
urllib3==1.26.12
urllib3==2.2.3
urllib3==2.3.0
vidgear==0.3.3
virtualenv==20.30.0
wrapt==1.14.1
yacs==0.1.8
